Refactored code : ParquetBuilderProperties merged in ParquetUtils
